Abstract Class for Macro Clusterers (Offline Component)

Abstract class for all DSC Macro Clusterers which recluster micro-clusters offline into final clusters called macro-clusters.

Usage

DSC_Macro(...)

microToMacro(x, micro = NULL)

Arguments

...

further arguments.

x

a DSC_Macro object that also contains information about micro-clusters.

micro

A vector with micro-cluster ids. If NULL then the assignments for all micro-clusters in x are returned.

Value

A vector of the same length as micro with the macro-cluster ids.

Details

Data stream clustering algorithms typically consist of an online component that creates micro-clusters (implemented as DSC_Micro) and offline components which are used to recluster micro-clusters into final clusters called macro-clusters. The function recluster() is used to extract micro-clusters from a DSC_Micro and create macro-clusters with a DSC_Macro.

Available clustering methods can be found in the See Also section below.

microToMacro() returns the assignment of Micro-cluster IDs to Macro-cluster IDs.

For convenience, a DSC_Micro and DSC_Macro can be combined using DSC_TwoStage.

DSC_Macro cannot be instantiated.

See also

Author

Michael Hahsler