Abstract Class for Implementing R-based Clusterers

Abstract class for implementing R-based clusterers.

Usage

DSC_R(...)

# S3 method for class 'DSC_R'
update(
  object,
  dsd,
  n = 1L,
  verbose = FALSE,
  block = 10000L,
  return = c("nothing", "assignment", "model"),
  ...
)

Arguments

...

further arguments.

object

a DSC object.

dsd

a data stream object.

n

number of data points taken from the stream.

verbose

logical; show progress?

block

process blocks of data to improve speed.

return

a character string indicating what update returns. The default is "nothing". Other possible values depend on the DST. Examples are "model" and "assignment".

Details

DSC_R cannot be instantiated.

Implementing new Classes

To implement a new clusterer you need to create an S3 class with elements description and RObj. RObj needs to be a reference class with methods:

cluster(newdata, ...)

get_microclusters(...)

get_microweights(...)

get_macroclusters(...)

get_macroweights(...)

microToMacro(micro, ...)

See DSC for details and parameters.

DSC_R cannot be instantiated.

See also

Author

Michael Hahsler