Data Stream Generator for Dynamic Data Stream Benchmarks

A data stream generator that generates several dynamic streams indented to be benchmarks to compare data stream clustering algorithms. The benchmarks can be used to test if a clustering algorithm can follow moving clusters, and merging and separating clusters.

Usage

DSD_Benchmark(i = 1)

Arguments

i

integer; the number of the benchmark.

Value

Returns a DSD object.

Details

Currently available benchmarks are:

1: two tight clusters moving across the data space with noise and intersect in the middle.

2: two clusters are located in two corners of the data space. A third cluster moves between the two clusters forth and back.

The benchmarks are created using DSD_MG.

See also

Author

Michael Hahsler

Examples

stream <- DSD_Benchmark(i = 1)
get_points(stream, n = 5)
#>          X1        X2 .class
#> 1 0.1127555 0.1000563      2
#> 2 0.1179650 0.1111449      2
#> 3 0.1155370 0.9058588      1
#> 4 0.1090585 0.1057890      2
#> 5 0.1105024 0.8997594      1

if (FALSE) { # \dontrun{
stream <- DSD_Benchmark(i = 1)
animate_data(stream, n = 10000, horizon = 100, xlim = c(0, 1), ylim = c(0, 1))

stream <- DSD_Benchmark(i = 2)
animate_data(stream, n = 10000, horizon = 100, xlim = c(0, 1), ylim = c(0, 1))
} # }