Abstract Class for Data Stream Regressors

Abstract class for data stream regressors. More implementations can be found in package streamMOA.

Usage

DSRegressor(...)

Arguments

...

Further arguments.

See also

Author

Michael Hahsler

Examples

DSClassifier()
#> DSClassifier is an abstract class and cannot be instantiated!
#> 
#> Available subclasses in ‘package:stream’ are:
#> 	DSClassifier_SlidingWindow
#> 
#> To get more information in R Studio, type ‘DSClassifier_’ and hit the Tab key.