Conceptual Base Class for All Data Stream Mining Tasks

Conceptual base class for all data stream mining tasks.

Usage

DST(...)

description(x, ...)

get_model(x, ...)

Arguments

...

Further arguments.

x

an object of a concrete implementation of a DST.

Details

Base class for data stream mining tasks. Types of DST are

DSAggregate to aggregate data streams (e.g., with a sliding window).

DSC for data stream clustering.

DSClassifier classification for data streams.

DSRegressor regression for data streams.

DSOutlier outlier detection for data streams.

DSFP frequent pattern mining for data streams.

The common interface for all DST classes consists of

update() update the DST with data points.

description() a string describing the DST.

get_model() returns the DST's current model (often as a data.frame or a R model object).

predict() use the learned DST model to make predictions.

and the methods in the Methods Section below.

See also

Author

Michael Hahsler

Examples

DST()
#> DST is an abstract class and cannot be instantiated!
#> Available subclasses are:
#> 	DSAggregate,
#> 	DSC,
#> 	DSClassifier,
#> 	DSD,
#> 	DSF,
#> 	DSFP,
#> 	DSOutlier,
#> 	DSRegressor