Simulate a jointly profiled scRNA-seq and scATAC-seq dataset in which several enhancer regulons (eRegulons) are contained.

Simulate a jointly profiled scRNA-seq and scATAC-seq dataset in which several enhancer regulons (eRegulons) are contained.

create_rna_atac(
  obj = NULL,
  ntfs = 5,
  ngenes = 50,
  ncells = 100,
  all.genes = 1000,
  all.enhs = 3000,
  all.cells = 1000,
  org = "hg38",
  atac.assay = "ATAC",
  gene.links = 2,
  multi.factor = 100,
  distance = 5e+05,
  url.link = "https://figshare.com/ndownloader/files/38794185"
)

Arguments

obj

A Seurat object used as the prototype to generate simulated dataset.

ntfs

The number of eRegulons (TFs) to include in the simulated dataset.

ngenes

The average number of genes in an eRegulon.

ncells

The average number of cells in an eRegulon.

all.genes

the number of genes in the simulated Seurat object.

all.enhs

The number of enhancers in the simulated Seurat object.

all.cells

The number of cells in the simulated Seurat object.

org

The organism, hg38 by default.

atac.assay

The scATAC-seq assay.

gene.links

The average number of enhancers linked to each gene in an eRegulon.

multi.factor

The value to be multiplied by the eRegulon genes

distance

The maximum distance between a gene and its linked enhancers.

Value

Returns a list composed of an eRegulon list and a Seurat object. each item of the simulated eRegulon contains the following attributes:

TF: The TF of the eRegulon.

genes: Genes of the eRegulon.

peaks: Enhancers of the eRegulon.

cells: Cells where the eRegulon is active.

References

Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022. Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data. bioRxiv, pp.2022-12.

Castro-Mondragon, J. A., Riudavets-Puig, R., Rauluseviciute, I., Berhanu Lemma, R., Turchi, L., Blanc-Mathieu, R., ... & Mathelier, A. (2022). JASPAR 2022: the 9th release of the open-access database of transcription factor binding profiles. Nucleic acids research, 50(D1), D165-D173.

Garcia-Alonso, L., Holland, C. H., Ibrahim, M. M., Turei, D., & Saez-Rodriguez, J. (2019). Benchmark and integration of resources for the estimation of human transcription factor activities. Genome research, 29(8), 1363-1375.