Plot an enhancer-driven gene regulatory network (eGRN)

Plot an enhancer-driven gene regulatory network (eGRN)

plot_eGRN(
  en.grn = NULL,
  obj = NULL,
  peak.assay = "ATAC",
  n.links = 20,
  title = "temp",
  path = "./",
  format = ".png",
  n.TFs = 15,
  layout = "fr",
  max.links = 5000,
  ifSaveRda = FALSE,
  TFs = NULL
)

Arguments

en.grn

The enhancer gene regulatory network (eGRN) composed of GRanges, cell type, and cells

obj

A Seurat object

peak.assay

The scATAC-seq assay

n.links

The cutoff of the number of TF-enhancer-gene relations for each TF to showcase. We retain the top-ranked relations in decreasing order of the number of cells, where the genes including the ones encoding the TFs are expressed, while the enhancers are accessible, 50 by default

title

The title of the eGRN to be plotted, "temp" by default

path

The path to save the image file, "./" by default

format

The format of the image file, ".png" by default

n.TFs

The maximum number of top-ranked TFs to generate eGRN plot, 10 by default

layout

The layout for plotting networks, "fr" by default

ifSaveRda

Whether save the igraph objects in a qsdave file

TFs

The list of TFs for plotting