stream: Infrastructure for Data Stream Mining

A framework for data stream modeling and associated data mining tasks such as clustering and classification. The development of this package was supported in part by NSF IIS-0948893, NSF CMMI 1728612, and NIH R21HG005912. Hahsler et al (2017) doi:10.18637/jss.v076.i14 .

Key Classes

DSD Data stream data and DSF data stream filter to transform data streams in flight.

DST Data stream mining tasks:

DSAggregate Data stream aggregator

DSC Data stream clustering

DSClassifier Data stream classifiers

DSOutlier Data stream outlier detectors

DSRegressor Data stream regressors

See also

Author

Maintainer: Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Authors:

Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Other contributors:

Matthew Bolaños [contributor]

John Forrest [contributor]

Matthias Carnein [contributor]

Dennis Assenmacher [contributor]

Dalibor Krleža [contributor]